+++ NEWS.md
@@ -1,3 +1,5 @@
+# isocountry (development version)
+
 # isocountry 0.6.0
 
 * Update the `isocurrency` data to the latest version (2026-02)